+++ demos/manual/src/main/java/is/codion/framework/demos/manual/common/demo/ApplicationPanel.java
@@ -22,7 +22,7 @@ import is.codion.common.format.LocaleDateTimePattern;
 import is.codion.common.item.Item;
 import is.codion.common.state.State;
 import is.codion.common.value.Value;
-import is.codion.swing.common.model.component.combobox.FilteredComboBoxModel.ItemFinder;
+import is.codion.swing.common.model.component.combobox.FilterComboBoxModel.ItemFinder;
 import is.codion.swing.common.model.component.combobox.ItemComboBoxModel;
 import is.codion.swing.common.ui.Sizes;
 import is.codion.swing.common.ui.component.Components;