+++ framework/model/src/main/java/is/codion/framework/model/EntityTableModel.java
@@ -21,7 +21,7 @@ package is.codion.framework.model;
 import is.codion.common.Configuration;
 import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.event.EventObserver;
-import is.codion.common.model.FilteredModel;
+import is.codion.common.model.FilterModel;
 import is.codion.common.model.table.TableSelectionModel;
 import is.codion.common.property.PropertyValue;
 import is.codion.common.state.State;
@@ -46,7 +46,7 @@ import java.util.Optional;
  * Specifies a table model containing {@link Entity} instances.
  * @param <E> the type of {@link EntityEditModel} used by this {@link EntityTableModel}
  */
-public interface EntityTableModel<E extends EntityEditModel> extends FilteredModel<Entity> {
+public interface EntityTableModel<E extends EntityEditModel> extends FilterModel<Entity> {
 
 	/**
 	 * Specifies the default action a table model takes when entities are inserted via its edit model.