+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/package-info.java
@@ -18,6 +18,6 @@
  */
 /**
  * Package configuration values:<br>
- * {@link is.codion.swing.common.model.component.combobox.FilteredComboBoxModel#COMBO_BOX_NULL_CAPTION}
+ * {@link is.codion.swing.common.model.component.combobox.FilterComboBoxModel#COMBO_BOX_NULL_CAPTION}
  */
 package is.codion.swing.common.model.component.combobox;