+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityConditionModelFactory.java
@@ -23,7 +23,7 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.framework.model.EntityConditionModelFactory;
-import is.codion.swing.common.model.component.combobox.FilteredComboBoxModel;
+import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
 import is.codion.swing.framework.model.component.EntityComboBoxModel;
 
 import java.util.Optional;
@@ -62,7 +62,7 @@ public class SwingEntityConditionModelFactory extends EntityConditionModelFactor
 	 */
 	protected EntityComboBoxModel createComboBoxModel(ForeignKey foreignKey) {
 		EntityComboBoxModel comboBoxModel = new EntityComboBoxModel(foreignKey.referencedType(), connectionProvider());
-		comboBoxModel.setNullCaption(FilteredComboBoxModel.COMBO_BOX_NULL_CAPTION.get());
+		comboBoxModel.setNullCaption(FilterComboBoxModel.COMBO_BOX_NULL_CAPTION.get());
 
 		return comboBoxModel;
 	}