+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -34,7 +34,7 @@ import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.i18n.FrameworkMessages;
 import is.codion.framework.model.EntitySearchModel;
-import is.codion.swing.common.model.component.combobox.FilteredComboBoxModel;
+import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
 import is.codion.swing.common.model.component.text.DocumentAdapter;
 import is.codion.swing.common.model.worker.ProgressWorker;
 import is.codion.swing.common.ui.Cursors;
@@ -634,7 +634,7 @@ public final class EntitySearchField extends HintTextField {
 		private static JPanel createSearchColumnPanel(EntitySearchModel searchModel) {
 			CardLayout cardLayout = new CardLayout(5, 5);
 			PanelBuilder columnBasePanelBuilder = panel(cardLayout);
-			FilteredComboBoxModel<Item<Column<String>>> columnComboBoxModel = new FilteredComboBoxModel<>();
+			FilterComboBoxModel<Item<Column<String>>> columnComboBoxModel = new FilterComboBoxModel<>();
 			EntityDefinition definition = searchModel.connectionProvider().entities().definition(searchModel.entityType());
 			for (Map.Entry<Column<String>, EntitySearchModel.Settings> entry : searchModel.settings().entrySet()) {
 				columnComboBoxModel.add(Item.item(entry.getKey(), definition.columns().definition(entry.getKey()).caption()));